RavenDB-18645 - GET /databases/*/collections/docs
