+++ src/Raven.Server/Documents/Handlers/CollectionsHandler.cs
@@ -1,15 +1,9 @@
-﻿using System.Diagnostics;
-using System.Threading.Tasks;
-using Raven.Client.Documents.Operations;
+﻿using System.Threading.Tasks;
 using Raven.Server.Documents.Handlers.Processors.Collections;
-using Raven.Server.Documents.Sharding.Commands;
 using Raven.Server.Documents.Sharding.Operations;
-using Raven.Server.Json;
-using Raven.Server.NotificationCenter.Notifications.Details;
 using Raven.Server.Routing;
 using Raven.Server.ServerWide.Context;
 using Sparrow.Json;
-using Sparrow.Json.Parsing;
 
 namespace Raven.Server.Documents.Handlers
 {
@@ -32,26 +26,8 @@ namespace Raven.Server.Documents.Handlers
         [RavenAction("/databases/*/collections/docs", "GET", AuthorizationStatus.ValidUser, EndpointType.Read)]
         public async Task GetCollectionDocuments()
         {
-            using (ContextPool.AllocateOperationContext(out DocumentsOperationContext context))
-            using (context.OpenReadTransaction())
-            {
-                var sw = Stopwatch.StartNew();
-                var pageSize = GetPageSize();
-                var documents = Database.DocumentsStorage.GetDocumentsInReverseEtagOrder(context, GetStringQueryString("name"), GetStart(), pageSize);
-
-                long numberOfResults;
-                long totalDocumentsSizeInBytes;
-                using (var token = CreateOperationToken())
-                await using (var writer = new AsyncBlittableJsonTextWriter(context, ResponseBodyStream()))
-                {
-                    writer.WriteStartObject();
-                    writer.WritePropertyName("Results");
-                    (numberOfResults, totalDocumentsSizeInBytes) = await writer.WriteDocumentsAsync(context, documents, metadataOnly: false, token.Token);
-                    writer.WriteEndObject();
-                }
-
-                AddPagingPerformanceHint(PagingOperationType.Documents, "Collection", HttpContext.Request.QueryString.Value, numberOfResults, pageSize, sw.ElapsedMilliseconds, totalDocumentsSizeInBytes);
-            }
+            using (var processor = new CollectionsHandlerProcessorForGetCollectionDocuments(this))
+                await processor.ExecuteAsync();
         }
 
         [RavenAction("/databases/*/collections/last-change-vector", "GET", AuthorizationStatus.ValidUser, EndpointType.Read)]
